app/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmtplib==2.0.2
anyio==3.7.0
APScheduler==3.10.1
async-timeout==4.0.2
asyncio-redis==0.16.0
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
boto3==1.26.147
botocore==1.29.147
broadcaster==0.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
config==0.5.1
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.96.0
fastapi-mail==1.2.8
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
http-ece==1.1.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.3
pandas==2.0.2
passlib==1.7.4
Pillow==9.5.0
psycopg2-binary==2.9.6
py-vapid==1.9.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
PyMySQL
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
pywebpush==1.14.0
redis==4.5.5
requests==2.31.0
rq==1.15.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
twilio==8.2.2
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
uuid==1.30
uvicorn==0.22.0
websockets==11.0.3
xattr==0.10.1
xlrd==2.0.1
yarl==1.9.2
